# contracts/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Correlation Id",
            "type": "string"
          },
          "error_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Error Version",
            "type": "string"
          },
          "message": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "correlation_id"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
      "AssetImageModel": {
        "additionalProperties": false,
        "properties": {
          "alt_text": {
            "title": "Alt Text",
            "type": "string"
          },
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "kind": {
            "const": "asset_image",
            "default": "asset_image",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "alt_text"
        ],
        "title": "AssetImageModel",
        "type": "object"
      },
      "BulletListModel": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ListItemModel"
            },
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "const": "bullet_list",
            "default": "bullet_list",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "title": "BulletListModel",
        "type": "object"
      },
      "CodeBlockModel": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "kind": {
            "const": "code_block",
            "default": "code_block",
            "title": "Kind",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "required": [
          "code",
          "language"
        ],
        "title": "CodeBlockModel",
        "type": "object"
      },
      "EducationalSourceModel": {
        "additionalProperties": false,
        "properties": {
          "license_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Note"
          },
          "locator": {
            "title": "Locator",
            "type": "string"
          },
          "publisher": {
            "title": "Publisher",
            "type": "string"
          },
          "reviewed_on": {
            "format": "date",
            "title": "Reviewed On",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "source_id",
          "title",
          "publisher",
          "locator",
          "reviewed_on",
          "license_note"
        ],
        "title": "EducationalSourceModel",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HardBreakModel": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "hard_break",
            "default": "hard_break",
            "title": "Kind",
            "type": "string"
          }
        },
        "title": "HardBreakModel",
        "type": "object"
      },
      "HeadingModel": {
        "additionalProperties": false,
        "properties": {
          "children": {
            "items": {
              "discriminator": {
                "mapping": {
                  "asset_image": "#/components/schemas/AssetImageModel",
                  "code": "#/components/schemas/InlineCodeModel",
                  "emphasis": "#/components/schemas/EmphasisModel",
                  "hard_break": "#/components/schemas/HardBreakModel",
                  "soft_break": "#/components/schemas/SoftBreakModel",
                  "source_link": "#/components/schemas/SourceLinkModel",
                  "strong": "#/components/schemas/StrongModel",
                  "text": "#/components/schemas/InlineTextModel"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/InlineTextModel"
                },
                {
                  "$ref": "#/components/schemas/InlineCodeModel"
                },
                {
                  "$ref": "#/components/schemas/SoftBreakModel"
                },
                {
                  "$ref": "#/components/schemas/HardBreakModel"
                },
                {
                  "$ref": "#/components/schemas/EmphasisModel"
                },
                {
                  "$ref": "#/components/schemas/StrongModel"
                },
                {
                  "$ref": "#/components/schemas/SourceLinkModel"
                },
                {
                  "$ref": "#/components/schemas/AssetImageModel"
                }
              ]
            },
            "title": "Children",
            "type": "array"
          },
          "kind": {
            "const": "heading",
            "default": "heading",
            "title": "Kind",
            "type": "string"
          },
          "level": {
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Level",
            "type": "integer"
          }
        },
        "required": [
          "level",
          "children"
        ],
        "title": "HeadingModel",
        "type": "object"
      },
...
      "InlineCodeModel": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "code",
            "default": "code",
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "InlineCodeModel",
        "type": "object"
      },
      "InlineTextModel": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "text",
            "default": "text",
            "title": "Kind",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "InlineTextModel",
        "type": "object"
      },
      "LessonReadingResponseModel": {
        "additionalProperties": false,
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/PassiveAssetModel"
            },
            "title": "Assets",
            "type": "array"
          },
          "body": {
            "items": {
              "discriminator": {
                "mapping": {
                  "bullet_list": "#/components/schemas/BulletListModel",
                  "code_block": "#/components/schemas/CodeBlockModel",
                  "heading": "#/components/schemas/HeadingModel",
                  "ordered_list": "#/components/schemas/OrderedListModel",
                  "paragraph": "#/components/schemas/ParagraphModel",
                  "thematic_break": "#/components/schemas/ThematicBreakModel"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/HeadingModel"
                },
                {
                  "$ref": "#/components/schemas/ParagraphModel"
                },
                {
                  "$ref": "#/components/schemas/CodeBlockModel"
                },
                {
                  "$ref": "#/components/schemas/ThematicBreakModel"
                },
                {
                  "$ref": "#/components/schemas/BulletListModel"
                },
                {
                  "$ref": "#/components/schemas/OrderedListModel"
                }
              ]
            },
            "title": "Body",
            "type": "array"
          },
          "objectives": {
            "items": {
              "type": "string"
            },
            "title": "Objectives",
            "type": "array"
          },
          "package_digest": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Package Digest",
            "type": "string"
          },
          "package_id": {
            "title": "Package Id",
            "type": "string"
          },
          "package_version": {
            "title": "Package Version",
            "type": "string"
          },
          "placement_id": {
            "title": "Placement Id",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/PublicationProvenanceModel"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/EducationalSourceModel"
            },
            "title": "Sources",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "placement_id",
          "package_id",
          "package_version",
          "package_digest",
          "title",
          "objectives",
          "body",
          "sources",
          "assets",
          "provenance"
        ],
        "title": "LessonReadingResponseModel",
        "type": "object"
      },
      "ListItemModel": {
        "additionalProperties": false,
        "properties": {
          "blocks": {
            "items": {
              "discriminator": {
                "mapping": {
                  "bullet_list": "#/components/schemas/BulletListModel",
                  "code_block": "#/components/schemas/CodeBlockModel",
                  "heading": "#/components/schemas/HeadingModel",
                  "ordered_list": "#/components/schemas/OrderedListModel",
                  "paragraph": "#/components/schemas/ParagraphModel",
                  "thematic_break": "#/components/schemas/ThematicBreakModel"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/HeadingModel"
                },
                {
                  "$ref": "#/components/schemas/ParagraphModel"
                },
                {
                  "$ref": "#/components/schemas/CodeBlockModel"
                },
                {
                  "$ref": "#/components/schemas/ThematicBreakModel"
                },
                {
                  "$ref": "#/components/schemas/BulletListModel"
                },
                {
                  "$ref": "#/components/schemas/OrderedListModel"
                }
              ]
            },
            "title": "Blocks",
            "type": "array"
          }
        },
        "required": [
          "blocks"
        ],
        "title": "ListItemModel",
        "type": "object"
      },
      "OrderedListModel": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ListItemModel"
            },
            "title": "Items",
            "type": "array"
          },
          "kind": {
            "const": "ordered_list",
            "default": "ordered_list",
            "title": "Kind",
            "type": "string"
          },
          "start": {
            "minimum": 1.0,
            "title": "Start",
            "type": "integer"
          }
        },
        "required": [
          "start",
          "items"
        ],
        "title": "OrderedListModel",
        "type": "object"
      },
      "ParagraphModel": {
        "additionalProperties": false,
        "properties": {
          "children": {
            "items": {
              "discriminator": {
                "mapping": {
                  "asset_image": "#/components/schemas/AssetImageModel",
                  "code": "#/components/schemas/InlineCodeModel",
                  "emphasis": "#/components/schemas/EmphasisModel",
                  "hard_break": "#/components/schemas/HardBreakModel",
                  "soft_break": "#/components/schemas/SoftBreakModel",
                  "source_link": "#/components/schemas/SourceLinkModel",
                  "strong": "#/components/schemas/StrongModel",
                  "text": "#/components/schemas/InlineTextModel"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/InlineTextModel"
                },
                {
                  "$ref": "#/components/schemas/InlineCodeModel"
                },
                {
                  "$ref": "#/components/schemas/SoftBreakModel"
                },
                {
                  "$ref": "#/components/schemas/HardBreakModel"
                },
                {
                  "$ref": "#/components/schemas/EmphasisModel"
                },
                {
                  "$ref": "#/components/schemas/StrongModel"
                },
                {
                  "$ref": "#/components/schemas/SourceLinkModel"
                },
                {
                  "$ref": "#/components/schemas/AssetImageModel"
                }
              ]
            },
            "title": "Children",
            "type": "array"
          },
          "kind": {
            "const": "paragraph",
            "default": "paragraph",
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "children"
        ],
        "title": "ParagraphModel",
        "type": "object"
      },
      "PassiveAssetModel": {
        "additionalProperties": false,
        "properties": {
          "alt_text": {
            "title": "Alt Text",
            "type": "string"
          },
          "asset_id": {
            "title": "Asset Id",
            "type": "string"
          },
          "locator": {
            "title": "Locator",
            "type": "string"
          },
          "media_type": {
            "enum": [
              "image/png",
              "image/jpeg",
              "image/webp"
            ],
            "title": "Media Type",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "asset_id",
          "media_type",
          "sha256",
          "alt_text",
          "locator"
        ],
        "title": "PassiveAssetModel",
        "type": "object"
      },
      "PublicationProvenanceModel": {
        "additionalProperties": false,
        "properties": {
          "content_reviewed_on": {
            "format": "date",
            "title": "Content Reviewed On",
            "type": "string"
          },
          "educational_use_notice": {
            "title": "Educational Use Notice",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "published_by": {
            "title": "Published By",
            "type": "string"
          }
        },
        "required": [
          "published_by",
          "published_at",
          "content_reviewed_on",
          "educational_use_notice"
        ],
        "title": "PublicationProvenanceModel",
        "type": "object"
      },
      "SessionBootstrapRequestModel": {
        "additionalProperties": false,
        "properties": {
          "limitation_acknowledged": {
            "const": true,
            "title": "Limitation Acknowledged",
            "type": "boolean"
          }
        },
        "required": [
          "limitation_acknowledged"
        ],
        "title": "SessionBootstrapRequestModel",
        "type": "object"
      },
      "SessionBootstrapResponseModel": {
        "additionalProperties": false,
        "properties": {
          "authentication_method": {
            "const": "single_profile",
            "default": "single_profile",
            "title": "Authentication Method",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "status": {
            "const": "ready",
            "default": "ready",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "expires_at"
        ],
        "title": "SessionBootstrapResponseModel",
        "type": "object"
      },
      "SoftBreakModel": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "soft_break",
            "default": "soft_break",
            "title": "Kind",
            "type": "string"
          }
        },
        "title": "SoftBreakModel",
        "type": "object"
      },
...
      "ThematicBreakModel": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "thematic_break",
            "default": "thematic_break",
            "title": "Kind",
            "type": "string"
          }
        },
        "title": "ThematicBreakModel",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OpaqueSessionCookie": {
        "description": "Opaque server-side session cookie. The value is never an identity selector and is not readable by browser scripts.",
        "in": "cookie",
        "name": "financial_ai_academy_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Financial AI Academy API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/curriculum/placements/{placement_id}/lesson": {
      "get": {
        "operationId": "getPlacedLesson",
        "parameters": [
          {
            "in": "path",
            "name": "placement_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 3,
              "pattern": "^[a-z0-9][a-z0-9._-]{2,63}$",
              "title": "Placement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonReadingResponseModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unavailable"
          }
        },
        "security": [
          {
            "OpaqueSessionCookie": []
          }
        ],
        "summary": "Get Placed Lesson",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/v1/session/single-profile": {
      "post": {
        "operationId": "bootstrapSingleProfileSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionBootstrapRequestModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionBootstrapResponseModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unavailable"
          }
        },
        "summary": "Bootstrap Single Profile Session",
        "tags": [
          "session"
        ]
      }
    },
...
  },
...
}